backend/requirements.txt
flask
flask-cors
flask-bcrypt
flask-jwt-extended
tensorflow
pillow
numpy
